Refactor into media component
