+++ src/plugins/index.ts
@@ -120,19 +120,6 @@ export const plugins: Plugin[] = [
           return `${supabaseUrl}/storage/v1/object/public/${bucket}/${filename}`
         },
       },
-      narrations: {
-        disablePayloadAccessControl: true,
-        generateFileURL: ({ filename }) => {
-          const supabaseUrl = process.env.NEXT_PUBLIC_SUPABASE_URL
-          const bucket = process.env.S3_BUCKET
-
-          if (!supabaseUrl || !bucket) {
-            return `/media/${filename}`
-          }
-
-          return `${supabaseUrl}/storage/v1/object/public/${bucket}/${filename}`
-        },
-      },
     },
     bucket: process.env.S3_BUCKET || "",
     config: {
